+++ FixtureController.php
@@ -293,8 +293,9 @@ class FixtureController extends \yii\console\controllers\FixtureController
     /**
      * Notifies user that given fixtures template files were not found.
      * @param array $templatesNames
+     * @since 2.0.4
      */
-    private function notifyNotFoundTemplates($templatesNames)
+    protected function notifyNotFoundTemplates($templatesNames)
     {
         $this->stdout("The following fixtures templates were NOT found:\n\n", Console::FG_RED);
 
@@ -307,8 +308,9 @@ class FixtureController extends \yii\console\controllers\FixtureController
 
     /**
      * Notifies user that there was not found any files matching given input conditions.
+     * @since 2.0.4
      */
-    private function notifyNoTemplatesFound()
+    protected function notifyNoTemplatesFound()
     {
         $this->stdout("No fixtures template files matching input conditions were found under the path:\n\n", Console::FG_RED);
         $this->stdout("\t " . Yii::getAlias($this->templatePath) . " \n\n", Console::FG_GREEN);
@@ -317,8 +319,9 @@ class FixtureController extends \yii\console\controllers\FixtureController
     /**
      * Notifies user that given fixtures template files were generated.
      * @param array $templatesNames
+     * @since 2.0.4
      */
-    private function notifyTemplatesGenerated($templatesNames)
+    protected function notifyTemplatesGenerated($templatesNames)
     {
         $this->stdout("The following fixtures template files were generated:\n\n", Console::FG_YELLOW);
 
@@ -329,7 +332,12 @@ class FixtureController extends \yii\console\controllers\FixtureController
         $this->stdout("\n");
     }
 
-    private function notifyTemplatesCanBeGenerated($templatesNames)
+    /**
+     * Notifies user about templates which could be generated.
+     * @param array $templatesNames
+     * @since 2.0.4
+     */
+    protected function notifyTemplatesCanBeGenerated($templatesNames)
     {
         $this->stdout("Template files path: ", Console::FG_YELLOW);
         $this->stdout(Yii::getAlias($this->templatePath) . "\n\n", Console::FG_GREEN);
@@ -346,8 +354,9 @@ class FixtureController extends \yii\console\controllers\FixtureController
      * by the given parameter.
      * @param array $templatesNames template file names to search. If empty then all files will be searched.
      * @return array
+     * @since 2.0.4
      */
-    private function findTemplatesFiles(array $templatesNames = [])
+    protected function findTemplatesFiles(array $templatesNames = [])
     {
         $findAll = ($templatesNames == []);
 
@@ -366,9 +375,11 @@ class FixtureController extends \yii\console\controllers\FixtureController
         $foundTemplates = [];
 
         foreach ($files as $fileName) {
-            $relativeName = str_replace(Yii::getAlias($this->templatePath) . '/', "", $fileName);   // strip templatePath from current template's full path
+            // strip templatePath from current template's full path
+            $relativeName = str_replace(Yii::getAlias($this->templatePath) . '/', "", $fileName);
             $relativeDir = dirname($relativeName) == '.' ? '' : dirname($relativeName) . '/';
-            $relativeName = $relativeDir . basename($relativeName,'.php');          // strip extension
+            // strip extension
+            $relativeName = $relativeDir . basename($relativeName,'.php');
             $foundTemplates[] = $relativeName;
         }
 
@@ -449,10 +460,13 @@ class FixtureController extends \yii\console\controllers\FixtureController
 
         $content = $this->exportFixtures($fixtures);
 
-        $dataFile = $fixtureDataPath . '/'. $templateName . '.php'; // data file full path
-        $dataFileDir = dirname($dataFile);                          // data file directory
-        if(!file_exists($dataFileDir)) {                            // if dir doesn't exist -
-            FileHelper::createDirectory($dataFileDir);              // create it
+        // data file full path
+        $dataFile = $fixtureDataPath . '/'. $templateName . '.php';
+
+        // data file directory, create if it doesn't exist
+        $dataFileDir = dirname($dataFile);
+        if (!file_exists($dataFileDir)) {
+            FileHelper::createDirectory($dataFileDir);
         }
         file_put_contents($dataFile, $content);
     }